+++ src/main/java/net/silentchaos512/gear/block/NetherwoodFence.java
@@ -1,0 +1,23 @@
+package net.silentchaos512.gear.block;
+
+import net.minecraft.block.BlockState;
+import net.minecraft.block.FenceBlock;
+import net.minecraft.util.Direction;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockReader;
+
+public class NetherwoodFence extends FenceBlock {
+    public NetherwoodFence(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public int getFlammability(BlockState state, IBlockReader world, BlockPos pos, Direction face) {
+        return 0;
+    }
+
+    @Override
+    public boolean isFlammable(BlockState state, IBlockReader world, BlockPos pos, Direction face) {
+        return false;
+    }
+}