+++ src/main/java/net/silentchaos512/gear/data/client/ModItemModelProvider.java
@@ -51,7 +51,6 @@ public class ModItemModelProvider extends ItemModelProvider {
         // Blocks
         Registration.BLOCKS.getEntries().stream()
                 .map(RegistryObject::get)
-                .filter(block -> block.asItem() != Items.AIR)
                 .forEach(this::blockItemModel);
 
         ModelFile itemGenerated = getExistingFile(new ResourceLocation("item/generated"));
@@ -113,11 +112,22 @@ public class ModItemModelProvider extends ItemModelProvider {
     }
 
     private void blockItemModel(Block block) {
+        if (block == ModBlocks.FLAX_PLANT.get() || block == ModBlocks.WILD_FLAX_PLANT.get()) {
+            return;
+        }
+
+        SilentGear.LOGGER.debug("blockItemModel: {}, {}", block, block.asItem());
         if (block == ModBlocks.PHANTOM_LIGHT.get()) {
             builder(block, getExistingFile(mcLoc("item/generated")), "item/phantom_light");
         } else if (block == ModBlocks.NETHERWOOD_SAPLING.get() || block == ModBlocks.STONE_TORCH.get()) {
             builder(block, getExistingFile(mcLoc("item/generated")), "block/" + NameUtils.from(block).getPath());
-        } else {
+        } else if (block == ModBlocks.NETHERWOOD_FENCE.get()) {
+            withExistingParent("netherwood_fence", modLoc("block/netherwood_fence_inventory"));
+        } else if (block == ModBlocks.NETHERWOOD_DOOR.get()) {
+            builder(block, getExistingFile(mcLoc("item/generated")), "item/netherwood_door");
+        } else if (block == ModBlocks.NETHERWOOD_TRAPDOOR.get()) {
+            withExistingParent("netherwood_trapdoor", modLoc("block/netherwood_trapdoor_bottom"));
+        } else if (block.asItem() != Items.AIR) {
             String name = NameUtils.from(block).getPath();
             withExistingParent(name, modLoc("block/" + name));
         }