+++ src/main/java/net/silentchaos512/gear/init/ModBlocks.java
@@ -27,20 +27,33 @@ import java.util.function.Supplier;
 
 public final class ModBlocks {
     public static final BlockRegistryObject<CrimsonIronOre> CRIMSON_IRON_ORE = register("crimson_iron_ore", CrimsonIronOre::new);
-    public static final BlockRegistryObject<MetalBlock> CRIMSON_IRON_BLOCK = register("crimson_iron_block", () -> new MetalBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(3.0f, 6.0f).sound(SoundType.METAL)));
-    public static final BlockRegistryObject<MetalBlock> CRIMSON_STEEL_BLOCK = register("crimson_steel_block", () -> new MetalBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(3.0f, 6.0f).sound(SoundType.METAL)));
-    public static final BlockRegistryObject<MetalBlock> BLAZE_GOLD_BLOCK = register("blaze_gold_block", () -> new MetalBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(3.0f, 6.0f).sound(SoundType.METAL)));
-    public static final BlockRegistryObject<GraderBlock> MATERIAL_GRADER = register("material_grader", () -> new GraderBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(5, 30)));
+    public static final BlockRegistryObject<MetalBlock> CRIMSON_IRON_BLOCK = register("crimson_iron_block", () ->
+            new MetalBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(3.0f, 6.0f).sound(SoundType.METAL)));
+    public static final BlockRegistryObject<MetalBlock> CRIMSON_STEEL_BLOCK = register("crimson_steel_block", () ->
+            new MetalBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(3.0f, 6.0f).sound(SoundType.METAL)));
+    public static final BlockRegistryObject<MetalBlock> BLAZE_GOLD_BLOCK = register("blaze_gold_block", () ->
+            new MetalBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(3.0f, 6.0f).sound(SoundType.METAL)));
+    public static final BlockRegistryObject<GraderBlock> MATERIAL_GRADER = register("material_grader", () ->
+            new GraderBlock(Block.Properties.create(Material.IRON).hardnessAndResistance(5, 30)));
     public static final BlockRegistryObject<CraftingStationBlock> CRAFTING_STATION = register("crafting_station", CraftingStationBlock::new);
     public static final BlockRegistryObject<SalvagerBlock> SALVAGER = register("salvager", SalvagerBlock::new);
     public static final BlockRegistryObject<FlaxPlant> FLAX_PLANT = registerNoItem("flax_plant", () -> new FlaxPlant(false));
     public static final BlockRegistryObject<FlaxPlant> WILD_FLAX_PLANT = registerNoItem("wild_flax_plant", () -> new FlaxPlant(true));
     public static final BlockRegistryObject<StoneTorch> STONE_TORCH = register("stone_torch", StoneTorch::new, bro -> getStoneTorchItem());
     public static final BlockRegistryObject<StoneTorchWall> WALL_STONE_TORCH = registerNoItem("wall_stone_torch", StoneTorchWall::new);
-    public static final BlockRegistryObject<NetherwoodLog> NETHERWOOD_LOG = register("netherwood_log", NetherwoodLog::new);
+    public static final BlockRegistryObject<NetherwoodLog> NETHERWOOD_LOG = register("netherwood_log", () ->
+            new NetherwoodLog(false));
+    public static final BlockRegistryObject<NetherwoodLog> STRIPPED_NETHERWOOD_LOG = register("stripped_netherwood_log", () ->
+            new NetherwoodLog(true));
     public static final BlockRegistryObject<NetherwoodPlanks> NETHERWOOD_PLANKS = register("netherwood_planks", NetherwoodPlanks::new);
     public static final BlockRegistryObject<NetherwoodSlab> NETHERWOOD_SLAB = register("netherwood_slab", NetherwoodSlab::new);
     public static final BlockRegistryObject<NetherwoodStairs> NETHERWOOD_STAIRS = register("netherwood_stairs", NetherwoodStairs::new);
+    public static final BlockRegistryObject<NetherwoodFence> NETHERWOOD_FENCE = register("netherwood_fence", () ->
+            new NetherwoodFence(Block.Properties.create(Material.WOOD).hardnessAndResistance(2f, 3f).sound(SoundType.WOOD)));
+    public static final BlockRegistryObject<NetherwoodDoor> NETHERWOOD_DOOR = register("netherwood_door", () ->
+            new NetherwoodDoor(Block.Properties.create(Material.WOOD).hardnessAndResistance(3f).sound(SoundType.WOOD).notSolid()));
+    public static final BlockRegistryObject<NetherwoodTrapdoor> NETHERWOOD_TRAPDOOR = register("netherwood_trapdoor", () ->
+            new NetherwoodTrapdoor(Block.Properties.create(Material.WOOD).hardnessAndResistance(3f).sound(SoundType.WOOD).notSolid()));
     public static final BlockRegistryObject<NetherwoodLeaves> NETHERWOOD_LEAVES = register("netherwood_leaves", NetherwoodLeaves::new);
     public static final BlockRegistryObject<NetherwoodSapling> NETHERWOOD_SAPLING = register("netherwood_sapling", NetherwoodSapling::new);
     public static final BlockRegistryObject<FlowerPotBlock> POTTED_NETHERWOOD_SAPLING = registerNoItem("potted_netherwood_sapling", () -> makePottedPlant(NETHERWOOD_SAPLING));
@@ -53,7 +66,9 @@ public final class ModBlocks {
     @OnlyIn(Dist.CLIENT)
     public static void registerRenderTypes(FMLClientSetupEvent event) {
         RenderTypeLookup.setRenderLayer(FLAX_PLANT.get(), RenderType.getCutout());
+        RenderTypeLookup.setRenderLayer(NETHERWOOD_DOOR.get(), RenderType.getCutout());
         RenderTypeLookup.setRenderLayer(NETHERWOOD_SAPLING.get(), RenderType.getCutout());
+        RenderTypeLookup.setRenderLayer(NETHERWOOD_TRAPDOOR.get(), RenderType.getCutout());
         RenderTypeLookup.setRenderLayer(POTTED_NETHERWOOD_SAPLING.get(), RenderType.getCutout());
         RenderTypeLookup.setRenderLayer(STONE_TORCH.get(), RenderType.getCutout());
         RenderTypeLookup.setRenderLayer(WALL_STONE_TORCH.get(), RenderType.getCutout());