+++ CHANGELOG.md
@@ -4,7 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Missing netherwood blocks (stripped log, fence, trapdoor, and door) [#88]
 ### Changed
+- Most Netherwood textures updated [#88]
 - Blaze gold's color to be more orange
 
 ## [1.9.0] - 2020-07-09